Move src/interface.rs into src/metric.
